Update test environment and import it as a python package instead as a git submodule (#158)
* Update test environment usage to new API of the test environment
* Adapt module imports to the python package
* Add the python package to the Pipfile
* Remove test testBigConnectionSysTables, because it doesn't test UDFs and fails with Exasol 7.0
* Fix test generic/combinations.py for Exasol 7.0

+++ githooks/update_current_submodules_file.sh
@@ -14,15 +14,20 @@ echo -e "Checking submodules ${grey}(pre-commit hook)${no_color} "
 # .git/)
 ROOT_DIR=$(git rev-parse --show-cdup)
 
-SUBMODULES=$(grep path ${ROOT_DIR}.gitmodules | sed 's/^.*path = //' | sort)
+GITMODULES_FILE="${ROOT_DIR}.gitmodules"
 
-CURRENT_GITMODULES="$ROOT_DIR.current_gitmodules"
-if [ -f "$CURRENT_GITMODULES" ]
+if [ -f "$GITMODULES_FILE" ]
 then
-  rm "$CURRENT_GITMODULES"
+  SUBMODULES=$(grep path "$GITMODULES_FILE" | sed 's/^.*path = //' | sort)
+
+  CURRENT_GITMODULES="$ROOT_DIR.current_gitmodules"
+  if [ -f "$CURRENT_GITMODULES" ]
+  then
+    rm "$CURRENT_GITMODULES"
+  fi
+  for SUB in $SUBMODULES
+  do
+      git ls-files -s "$SUB" >> "$CURRENT_GITMODULES"
+  done
+  git add "$CURRENT_GITMODULES"
 fi
-for SUB in $SUBMODULES
-do
-    git ls-files -s "$SUB" >> "$CURRENT_GITMODULES"
-done
-git add "$CURRENT_GITMODULES"
